MyOrthotics 2.0
MyOrthotics 2.0
Bio inspired structures and human interaction are infinite source of
inspiration applied in wearables, rehabilitation, prosthetics and other
assistive devices.The process of making assistive devices has until
recent years been a long and expensive individualized process,
requiring detailed development and bespoke labor by physical therapists
and manufacturers. We will explore the digital craftsmanship, using
digital manufacturing (design of patterns, 3D Scanning, 3D Printing and
electronics) to produce an individual DIY solution for disabilities of
the hand and arm that will allow patients, therapists, and
practitioners to develop this assistive device in collaborative maker
spaces.
MyOrthosis 2.0 is an orthosis developed for workshops to introduce the
concept of assistive devices and it is based on a workflow from 3D
Scanning, 3D printing , cutting and electronics to explore rapid
construction for orthotics for first aid cases or temporary use.